growth_intelligence/requirements.txt
# Core orchestration
langgraph>=0.2.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langchain>=0.2.0

# Memory & vector store
mem0ai>=0.1.0
pinecone>=3.0.0

# Web scraping & search
firecrawl-py>=0.0.16
httpx>=0.27.0

# Social signals
asyncpraw>=7.7.0

# Deep parse
llama-parse>=0.4.0

# Structured output & validation
pydantic>=2.0.0

# PDF export
weasyprint>=62.0
jinja2>=3.1.0

# Observability
langfuse>=2.0.0

# Utilities
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
